move simple streamer to graphql module
